Title: 겹치지 않는 선분위의 두 점, Time: 200ms, Memory: 14MB, Status: Wrong Answer - Codetree

+++ 250611/겹치지 않는 선분위의 두 점/two-points-of-the-line-that-dont-overlap.java
@@ -21,7 +21,7 @@ class Point implements Comparable<Point> {
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int n = sc.nextInt();
+        long n = sc.nextInt();
         int m = sc.nextInt();
         Point[] points = new Point[m];
         for (int i = 0; i < m; i++) {
